+++ projects/fusion-docs/src/app/pages/docs/docs.component.ts
@@ -11,6 +11,12 @@ import {DocsLayoutService} from './docs-layout.service';
 import {VersionService} from '../../services/version/version.service';
 import {StyleVersion} from '@ironsource/fusion-ui/components/fusion-base';
 import {LayoutComponentConfiguration, LayoutHeaderComponentConfiguration} from '@ironsource/fusion-ui/components/layout/v2';
+import {NAVIGATION_MENU_MOCK} from '@ironsource/fusion-ui/components/navigation-menu/v4/stories/navigation-menu.mock';
+import {HeaderContent, LayoutConfiguration} from '@ironsource/fusion-ui/components/layout/v4/layout.entities';
+import {TopFilterIncludeExcludeComponent} from '@ironsource/fusion-ui/components/top-filter-include-exclude';
+import {FormControl} from '@angular/forms';
+import {DropdownOption} from '@ironsource/fusion-ui/components/dropdown-option';
+import {MOK_APPLICATIONS_ONE_LINE_OPTIONS} from '@ironsource/fusion-ui/components/dropdown/v3/stories/dropdown.mock';
 
 @Component({
     selector: 'fusion-docs',
@@ -19,6 +25,7 @@ import {LayoutComponentConfiguration, LayoutHeaderComponentConfiguration} from '
 })
 export class DocsComponent implements OnInit, OnDestroy {
     useNewLayout = true; // switch to new layout
+    useLayoutName = 'layoutV4'; // 'layoutV1', 'layoutV2'
     menuItems: MenuItem[] | SidebarMenuItem[] = this.useNewLayout ? MENU_ITEMS_V2 : MENU_ITEMS;
 
     // region for layout-v1 v1
@@ -38,6 +45,32 @@ export class DocsComponent implements OnInit, OnDestroy {
     layoutConfiguration$ = this.docsLayoutService.layoutConfig$;
     // endregion
 
+    // region for layout "v4"
+    appSelectFormControl = new FormControl(MOK_APPLICATIONS_ONE_LINE_OPTIONS.slice(1, 5)) as FormControl<DropdownOption[]>;
+    headerContent: HeaderContent = {
+        title: 'Dashboard',
+        actionComponent: TopFilterIncludeExcludeComponent,
+        actionData: {
+            placeholder: 'Select application',
+            formControl: this.appSelectFormControl,
+            title: 'Applications',
+            items: MOK_APPLICATIONS_ONE_LINE_OPTIONS
+        }
+    };
+    layoutConfiguration: LayoutConfiguration = {
+        navigationMenuItems: NAVIGATION_MENU_MOCK,
+        layoutUser: {
+            name: 'Jonny Kim',
+            email: 'jonny.kim@supercell.com'
+        }
+    };
+    // endregion
+
+    placeholder = 'Select application';
+    formControl = new FormControl(MOK_APPLICATIONS_ONE_LINE_OPTIONS.slice(1, 4));
+    title = 'Applications';
+    items = MOK_APPLICATIONS_ONE_LINE_OPTIONS;
+
     private onDestroy$ = new Subject<void>();
     private selectedVersion$ = this.versionService.styleVersion$;
 
@@ -47,6 +80,14 @@ export class DocsComponent implements OnInit, OnDestroy {
         // region set layout "v2"
         this.setLayoutConfigurationObservable();
         // endregion
+
+        this.formControl.valueChanges.pipe(takeUntil(this.onDestroy$)).subscribe(val => {
+            console.log('11111', val);
+        });
+
+        this.appSelectFormControl.valueChanges.pipe(takeUntil(this.onDestroy$)).subscribe(val => {
+            console.log(':::::', val);
+        });
 
         this.selectedVersion$.pipe(takeUntil(this.onDestroy$)).subscribe(value => {
             const menuItemsAll = JSON.parse(JSON.stringify(this.useNewLayout ? MENU_ITEMS_V2 : MENU_ITEMS));
@@ -131,6 +172,16 @@ export class DocsComponent implements OnInit, OnDestroy {
         if (menuItem.redirect) {
             location.href = menuItem.redirect;
         }
+    }
+
+    onMenuItemClick(menuItem: MenuItem) {
+        console.log('DOC onMenuItemClick>>>', menuItem);
+
+        this.headerContent = {...this.headerContent, title: menuItem.name};
+
+        // if (menuItem.redirect) {
+        //     location.href = menuItem.redirect;
+        // }
     }
 
     onLayoutUserLogout(user) {